relocate map component file to match form
