+++ src/DashboardUI/src/App.tsx
@@ -32,7 +32,7 @@ import WaterRightDetailsPage from './pages/WaterRightDetailsPage';
 import { ApplicationSubmitPage } from './pages/application/dashboard/ApplicationSubmitPage';
 import ApplicationReviewPage from './pages/application/review/ApplicationReviewPage';
 import { ApplicationReviewGuard } from './pages/application/review/ApplicationReviewGuard';
-import ApplicationReviewMap from './pages/application/review-map/ApplicationReviewMap';
+import ApplicationReviewMap from './pages/application/review/review-map/ApplicationReviewMap';
 import ApplicationReviewFormLayout from './pages/application/review/form/ApplicationReviewFormPage';
 
 import './App.scss';
